Create GPIO
In the user interface of this project, there are a total of four buttons, which are actually the making of four kinds of coffee. In the coffee machine, controlling the amount of coffee beans, milk consumption and water flow of different coffees is actually realized by controlling sensors and relays, while I simply control the GPIO pin first.